cover_buildup/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==2.2.19
django-filter==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
oauthlib==3.2.0
openpyxl==3.0.10
pandas==1.3.5
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1
